+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 
 * Fix a potential Lua error upon login [#230]
 * Fix "spell known" load condition in Classic Era and Cataclysm classic [#227]
+* Fix Lua error with the "form /stance" macro condition on Classic Era [#226]
 
 ## [1.16.4] - 2024-09-08
 
@@ -1651,6 +1652,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 
 [#230]: https://github.com/Snakybo/Clicked/issues/230
 [#227]: https://github.com/Snakybo/Clicked/issues/227
+[#226]: https://github.com/Snakybo/Clicked/issues/226
 [#224]: https://github.com/Snakybo/Clicked/issues/224
 [#223]: https://github.com/Snakybo/Clicked/issues/223
 [#217]: https://github.com/Snakybo/Clicked/issues/217